requirements.txt
coverage
django
pylint
pylint-django
python-dateutil
